+++ db/migrate/20160508135728_create_route_search_points.rb
@@ -1,8 +1,8 @@
 class CreateRouteSearchPoints < ActiveRecord::Migration
   def change
     create_table :route_search_points do |t|
-      t.decimal :longitude, default: 0.0
-      t.decimal :latitude, default: 0.0
+      t.float :longitude, default: 0.0
+      t.float :latitude, default: 0.0
 
       t.integer :route_id, default: 0
       t.timestamps null: false
